requirements.txt
Django==4.0.5
djangorestframework
markdown
django-filter
PyJWT
django-cors-headers
Pillow==9.1.1